requirements.txt
discord.py
asyncio
aiohttp
python-dateutil
dataset
typing
requests
bs4
pymediawiki
rule34